Refactor parse to use dedicated class to expressions
